src/main/java/kr/megaptera/assignment/application/DeleteCommentService.java
package kr.megaptera.assignment.application;

import kr.megaptera.assignment.domains.dto.CommentDto;
import kr.megaptera.assignment.domains.model.Comment;
import kr.megaptera.assignment.domains.model.CommentId;
import kr.megaptera.assignment.exceptions.CommentNotFound;
import kr.megaptera.assignment.repositories.CommentRepository;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class DeleteCommentService {

    private final CommentRepository commentRepository;

    public DeleteCommentService(CommentRepository commentRepository) {
        this.commentRepository = commentRepository;
    }

    public CommentDto deleteComment(String id) {
        Optional<Comment> check = commentRepository.findById(CommentId.of(id));
        if (check.isEmpty())
            throw new CommentNotFound();
        Comment comment = check.get();
        commentRepository.delete(comment);
        return new CommentDto(comment);
    }
}
